refactor: table component 분리
